Fix best path calculation
Currently "GetClosestPowerSocket" & "GetClosestUnreservedHibernationSpot" do not return the "closest" but the first one found.
This commit fixes the issue by going through all nodes and returning the one with the shortest distance.

Depending on the number of nodes present on the map, it could bit a bit slower but the difference should be negligible.
Also, it will allows Automatons to choose the quickest path to their destination.